Move runner into package
